Added a start_squash option to the event filter.
And used it to improve the 'Job History' display.

With start_squash enabled, the filter removes starts which
correspond to a finish, and places the elapsed time inside
the finish event.

Also made use of the utility classes from inside
the job.html template to avoid having to replicate
logic there.  Should do the same for the other templates.