Web: abbreviated long command names.
Added an abbr function which can be used as a filter in
Mako and applied it to command strings in page headings
and in the tables giving a list of cron jobs.
The full command is present in the span title, so should
be visible if the mouse is hovered over the abbreviated
version.