Templ, CSS: added drop-down menu to 'Crab logo'.
Hovering the mouse over the Crab H1 element should
now cause a menu to appear.  This provides access to
the Host/User Notifications page, and will allow
links to any future administrative pages to be accommodated.
Menu is implemented as a UL element, and has been tested
with Firefox and Chromium.