Store: corrected job indentification regression.
In commit debbd0dcd0 where the explicit SQL statements were removed
from the _check_job method and replaced with calls to _get_jobs,
the part which assigned job identifiers to previously unidentified
jobs lost its search condition restricting it to unidentified jobs.
This commit adds an option to the _get_jobs method allowing this
condition to be reinstated.  Also added a test to the store test
script to check that this is working as expected.