Web: use Font Awesome for icons.
Font Awesome is an iconic font supplied with a stylesheet which can
be used to apply the icons to HTML elements.  Using this stylesheet
ensures that the codepoints are correct for the corresponding
version of the font.

Applied an icon to the RSS link in the header bar, and to various
other links in the Crab web interface.

Added a note to the README.rst file describing how to install
the font into the Crab 'res' directory.