Setup: find and add additional files.
The templates, resources and documentation files are now
detected by the setup script and added as "data_files".

The Python modules are also now detected rather than
listed in the setup script.

Since the README was renamed to README.rst, it must now be
included explicitly in the MANIFEST.

Set version number to 0.1.0.