Web: added job event history navigation.
There are now options to select the number of events displayed,
to show all event types, and to step back through the history.

History navigation is done by date in order to maintain the
interface to the get_job_events method, and for simplicity
(no need to handle the various event ID numbers).
The template code generating the content of the job event history
has been moved to a separate file, which can be called
to generate just the table rows.  This is used when the
JavaScript requests the page, giving the barerows parameter.

Moved the squash_start parameter from the constructor
of CrabEventFilter to its call method.