Store: implementing file storage of job output.
This involved implementing the dummy methods of CrabStoreFile,
and updating the configuration file to allow an output store to be
configured.

Added the user-specified jobid (name) to the interface for reading
and writing job output to allow CrabStoreFile to use it as part
of the directory structure instead of the job number.