Util: added utilities directory.
This is a directory to house utilities for use in
administration of a Crab server.

Added scripts to transfer output and crontab records
to and from an output store.  To do this it was helpful
to allow the get_job_finishes method not to limit the number
of finish records returned.