Store: handle number vs named directories better.
When logging job output for an unnamed job, now check
whether a name is recorded for it, to allow the file store
to put the output in the correct directory.

When reading job output, if the file store cannot find
the file, it also checks the numeric directory in
case output was recorded before the job was given a
name.  It is of course preferable to assign a name
using the CRABID parameter to ensure that the output
goes in the correct directory from the start.