Util: add split_crab_vars function.
This parses CRAB variables from the start of a command string.
It is now used in importing crontabs and in crabsh.
Variables from the environment and those from the command
string are now merged by crabsh.

Also added a true_string function to check whether a string
should be taken to be a true value.  This avoids repeating the list
of false strings in various places.