Store: track crontab variables properly on import.
Now the CRAB variables in the crontab are placed in a dictionary,
and merged with variables set for each command.  This allows it
to handle CRABIGNORE environment variables in the crontab.

It also makes the handling of CRABID match that used by crabsh,
with the result that warnings can be given if a CRABID specified
by environment variable applies to multiple jobs.