Templ, CSS: add support for Font Awesome 3.
Updated the attribution in the header comment to match that now
suggested on the Font Awesome website.

With the previous version's CSS file, you could get away with
using the icon classes on any element to apply the icon to that
element.  However in this version the same classes now style
the whole element (not just its :before pseudo-element).
Therefore the Crab templates now apply the icon classes to
empty span elements.  (I used <span> instead of <i>, unlike the
examples on the Font Awesome website, because <i> doesn't
mean "icon".)