Web: make PyRSS2Gen dependency optional.
This was done because, following the introduction of email
notifications, RSS is no longer critical.

If it is not present, importing crab.web.rss raises an ImportError
which crabd now traps.  It then passes a new "options" dict
(for lack of a better name) to the CrabWeb constructor including
an entry "rss_enabled" indicating whether or not RSS is available.

Also tidied up the CrabWeb constructor by no longer passing the whole
config object.  The Crab server component constructors should only
be given the configuration parameters they actually read, rather
than having to deal with the configuration in its entirety.