JS: improve history handling on job history page.
The current history state is now updated as the user
manipulates the display via the form elements and
previous / last controls.  So when navigating back to
this page (e.g. using the browser back button after
having looked at a job output page), it should
load the history contents as previously viewed.

In addition, where supported, the current 'end date'
is read back from the history state while updating
the list based on the form elements, so that this
action also does not loose the current place in the
job history.