Templ, JS: offer coloring of ANSI output.
When a cron job output contains ANSI colors, then
a link will be displayed above the output text.  Clicking
this link should cause the output to be displayed in color.

This is implemented in the web browser, using the ansi_up
JavaScript library.  It has therefore been added to the README
file as an optional requirement.