Server: default timezone to notification timezone.
The CrabEventFilter default timezone is now set to the
timezone configured in the notification section of the
server configuration.  This is probably more likely
to be useful than always defaulting to UTC, but a second
configuration parameter could be added if a need
arises to configure the daily notification timezone
separately from the default display timezone for
notifications and the web interface.