Store: check for EEXIST when making directories.
Unlike "mkdir -p", os.makedirs doesn't permit the directory
to already exist.  There is an exist_ok parameter but it doesn't
exist in all versions of Python targetted by the Crab server,
so it can't be used here.

Instead catch the error, but ignore it if the error code
was EEXIST.