Templ, CSS, JS: add support for Font Awesome 4.
Updated the attribution in the header comment to give the URL now
specified on the Font Awesome website.

Font Awesome 4 changes the name of the "font" directory to "fonts"
and alters the CSS selectors, requiring changes to the classes used
to display the icons.  The change required for existing Crab
servers is described in the CHANGES.txt file.