Store: move start/finish logic to CrabStore class
The additional logic in the log_start and log_finish methods has been
moved to the main CrabStore class.  The CrabStoreDB class now has
simple "private" methods which simply perform the database record
insertions.  This is to make it easier to implement other store
classes -- they should not have to repeat this logic.