Server: add facility to import/export config
This adds a new module crab.server.io and crabd command line
parameters --import and --export.

Jobs, their configuration, notifications and raw cronabs are
handled by the new module.  It uses a JSON formatted file.
The purpose of the module is to allow a Crab service to be
migrated between different databases, as well as for backup
and restoration.