Service: set seconds to 0.0 in run_minutely calls
The notification service already had to set the seconds and microseconds
to zero.  The monitor service ought to have been doing this too since
it sets the missed job timer based on the given datetime.  Now the
CrabMinutely base class _check_minute method sets the seconds
and microseconds to zero when calling the run_minutely method,
which avoids the individual services having to do this.  The result
should be more accurate timing of job missed alarms.