Store: move cursor and error control to lock class
The CrabDBLock class now returns a database cursor from its
__enter__ method, avoiding the need for manual acquisition and
closing of cursors.  Also the __exit__ method now looks for
database errors and converts them to CrabError, removing the
need for try: blocks around each database routine.

The motivation for this is that it is possible for a cursor's
close method to raise an exception.  The standard contextlib.closing
context manager does not trap this exception, so a custom
context manager was required, and it was simplest to include this
functionality in the existing "lock" context manager.

Now all private database methods take a cursor object 'c' as their
first argument.