Database: allow jobfinish deletions to cascade
Now the foreign key constraint on the joboutput.finishid column
uses "ON DELETE CASCADE" so that, when an outputstore is not used
(so that job output is stored in the database), deleting a finish
record causes the associated output also to be deleted.  (And a
foreign key constraint error is avoided.)

This is to allow the creation of an automated cleaning routine to
limit the length of time for which event records are kept.

Update scripts have been added to the util directory for updating
existing databases.