Client: decode output immediately in crabsh
In commit 9ef8fa972848d585e372492b4084c60c842f2c29, decoding of
the job output to unicode was added for transmission to the Crab
server.  However this left the output in bytes (in Python 3)
so that it was printed incorrectly in event of a failure to notify
of the job finish, or usage of CRABECHO.  Now decode the output
immediately to avoid this problem on Python 3.