Client: improve handling of PID files in crabsh
Now write the PID of the crabsh process itself rather than the
PID of the child process into the PID file.  This means that the
file can be written on startup, rather than after waiting to
communicate with the Crab server and to start the sub-process.

The CRABPIDFILE option is now handled even when CRABIGNORE is set.
This means that crabsh can be used to apply PID file control to an
ignored job.