Util: process percent signs in crontabs
This updates the parse_crontab and write_crontab functions to handle
percent signs (escaped and unescaped) in cron job specifications.
Both functions process the "input" -- the portion of the line after
the first %-sign -- and deal with backslash-escaping of %-signs
not used as input separators.  This is to allow round-trip processing
of crontab lines to be tested -- at present Crab does not store the
job input.