Client: split client hostname at first dot
On some systems, socket.gethostname() returns a fully qualified domain
name.  This was not what was intended for the client.hostname configuration
parameter, so split the hostname at the first dot and take only the actual
name of the host.

For cases where the fully quailified domain name was actually desired,
a new configuration parameter client.use_fqdn has been added.  This
causes socket.getfqdn() to be used instead (and the result not split
at the first dot).  In some situations, socket.gethostname() returns
a fully qualified domain name but socket.getfqdn() does not -- in this
case the new option (client.use_fqdn=true) will not work and the client
hostname must be configured directly.