Client: add user and host name env. variables
This is to allow the client's user name and host name to be more easily
overridden (e.g. on systems which share a home directory so that it is
not convenient to do this via the user configuration files).

While normally the server address is called "host" and the client's name
is just called "hostname", the new environment variable for client name is
called "CRABCLIENTHOSTNAME" in order to make its role clear and reduce the
risk of confusion.