Email: include report summary in message subject
This adds a new report function (crab.report.summary.report_to_summary)
which produces a very short summary of a report -- currently just the titles
of the jobs having the most severe status, with limits on the number of
jobs to list and the length of each title.

Ideally this would be configurable via format specifiers in the subjects
given by the email section of the "crab.ini" file.  However for now the
summary is simply appended to the subject, in parentheses.