Server: skip import of jobs for other host/user
Now when saving a crontab, skip jobs for which the CRABCLIENTHOSTNAME
or CRABUSERNAME is set but does not match the host / user to which this
crontab is supposed to belong.

Note that the parse_crontab function now includes the whole crontab
line (as the "rule" dictionary entry) to allow CrabStore.save_crontab
to generate more useful warnings.