Server: add support for rotating log files
This adds two new command line options to crabd: --accesslog and
--errorlog.  If given, these specify where the corresponding
CherryPy logs are written.  The file rotation parameters can be
configured via the crabd.ini file.  The crabd-check script now
specifies these parameters rather than redirecting output from
crabd to a log file.

The error log combines the CherryPy "error" log (not just for errors)
with output from other loggers -- the log file handler is attached
to the root logger and CherryPy's logger propagates to it.  To keep
messages from CherryPy's "access" logger separate, the file logger
is attached directly to it an propagation is disabled.  If file
names are not specified, the previous "screen" behavior of CherryPy
should be replicated -- the access log will go to stdout and the error
log to stderr.

Note that CherryPy adds the time to the message text of error log
entries.  However to ensure that general log messages sent to the
error log file always include the time, it is added again in the
handler attached to the root logger.