Add support for using CherryPy's daemonizer plugin
This adds a "--daemon" option to "crabd" to have it set up the CherryPy
daemonizer.  The "crabd-check" script now uses this option rather than
simply running "crabd" in the background.

However the Crab services which run in separate threads (the monitor,
notification service and clean service) must now be started after
the daemonizing process by listening on the "start" channel of the
CherryPy bus.  The Crab store and notification sender are also started
in this way in case they acquire any resources which should wait
until after daemonization.  The PID file is also written afterwards
so that it contains the persistent PID.  The start up process should
be as follows (with priorities for the "start" channel):

    65: CherryPy deamonizer runs

    70: Crab facilities (store and notification sender) set up

    71: Crab service threads launched

    75: CherryPy web server starts

Since the Crab services are intended to be independent of CherryPy
(in case it is ever desired to run them separately) they have been left
unaltered but are each launched via a new intermediate "CrabPlugin" object
which receives the Crab store object (and notification sender if needed)
via the CherryPy bus, then constructs the service and starts it.

The Crab "server" and web interface classes already interact with
CherryPy, so they now subscribe to the CherryPy bus to get hold of the
store, monitor and other services when they become available.