Client: add server.comm_timeout config. parameter
The HTTPConnection is created with the "timeout" value as before
(but now read as a float instead of as an int).  After connection,
if "comm_timeout" is specified, attempt to adjust the connection's
socket object to the given timeout (also read as a float).

This is to allow for longer timeouts communicating with crabd
once a connection has been established.  Retries are not possible
at this stage since crabd may already be processing a message
sent to it.  The connection timeout (and retries) can be kept
reasonably short to avoid delaying the start of cron jobs.